Create Paging and Cash

+++ app/src/main/java/com/example/newsapp/ui/main/MainActivity.java
@@ -1,20 +1,22 @@
 package com.example.newsapp.ui.main;
 
-import androidx.annotation.RequiresApi;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.core.widget.NestedScrollView;
 import androidx.lifecycle.ViewModelProviders;
 import androidx.recyclerview.widget.DividerItemDecoration;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
+
 import android.content.Context;
 import android.content.Intent;
 import android.net.ConnectivityManager;
-import android.os.Build;
 import android.os.Bundle;
 import android.util.Log;
 import android.view.View;
 import android.widget.ProgressBar;
+import android.widget.Toast;
+import android.widget.Toolbar;
+
 import com.example.newsapp.App;
 import com.example.newsapp.R;
 import com.example.newsapp.models.Article;
@@ -31,12 +33,11 @@ public class MainActivity extends AppCompatActivity {
     private RecyclerView recyclerView;
     private NewsAdapter adapter;
     private List<Article> list = new ArrayList<>();
-    private ProgressBar progressBar;
+    private ProgressBar isLoading, progress;
     private NestedScrollView nestedScrollView;
-    private Integer pageSize = 10;
-    private Integer page = 1;
+    private int pageSize = 10;
+    private int page = 1;
 
-    @RequiresApi(api = Build.VERSION_CODES.M)
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -44,41 +45,42 @@ public class MainActivity extends AppCompatActivity {
         initViews();
         createRecycler();
         getDataFromLiveData();
-        mViewModel.setIsLoading();
         listeners();
-
     }
 
     private void getDataFromLiveData() {
         ConnectivityManager cm = (ConnectivityManager) getSystemService(Context.CONNECTIVITY_SERVICE);
         if (cm.getActiveNetworkInfo() != null && cm.getActiveNetworkInfo().isConnected()) {
-            App.newsDataBase.newsDao().deleteAll();
-            mViewModel.receiveData(pageSize, page);
+
+            mViewModel.receiveData(page, pageSize);
+            if (App.newsDataBase.newsDao().getAll() != null) App.newsDataBase.newsDao().deleteAll();
+
             mViewModel.news.observe(this, result -> {
                 App.newsDataBase.newsDao().insert(result);
-                list.addAll(App.newsDataBase.newsDao().getAll());
+                list.addAll(result);
                 adapter.updateAdapter(list);
+                progress.setVisibility(View.GONE);
             });
-            Log.d("ololo", "Internet is connected");
         } else {
             mViewModel.newsData.observe(this, articles -> {
                 if (articles != null) {
+                    list.addAll(articles);
                     adapter.updateAdapter(articles);
-                    list = articles;
+                    isLoading.setVisibility(View.GONE);
                 }
             });
-            Log.d("ololo", "Internet is disconected");
         }
         mViewModel.isLoading.observe(this, aBoolean -> {
-            if (aBoolean) progressBar.setVisibility(View.GONE);
+            if (aBoolean) isLoading.setVisibility(View.GONE);
         });
     }
 
     private void initViews() {
         mViewModel = ViewModelProviders.of(this).get(MainViewModel.class);
         recyclerView = findViewById(R.id.recycler_view);
-        progressBar = findViewById(R.id.progress_bar);
+        isLoading = findViewById(R.id.progress_bar);
         nestedScrollView = findViewById(R.id.nested_scroll);
+        progress = findViewById(R.id.progress_bar_down);
     }
 
     private void listeners() {
@@ -86,29 +88,26 @@ public class MainActivity extends AppCompatActivity {
             startActivity(new Intent(this, DetailsActivity.class).putExtra("pos", list.get(pos)));
         });
 
-        nestedScrollView.setOnScrollChangeListener(new NestedScrollView.OnScrollChangeListener() {
-            @Override
-            public void onScrollChange(NestedScrollView v, int scrollX, int scrollY, int oldScrollX, int oldScrollY) {
-                if (scrollY == v.getChildAt(0).getMeasuredHeight() - v.getMeasuredHeight());
-                page++;
-                pageSize = 10;
-                mViewModel.receiveData(pageSize, page);
-                Log.d("ololo", " " + page);
-            }
-        });
+        nestedScrollView.setOnScrollChangeListener((NestedScrollView.OnScrollChangeListener)
+                (v, scrollX, scrollY, oldScrollX, oldScrollY) -> {
+                    if (scrollY == v.getChildAt(0).getMeasuredHeight() - v.getMeasuredHeight()){
+                        if (pageSize >= list.size()) {
+                            page++;
+                            pageSize = pageSize + 10;
+                            progress.setVisibility(View.VISIBLE);
+                            mViewModel.receiveData(page, pageSize);
+                            Log.d("ololo", " page " + page + "pageSize" + pageSize);
+                        } else {
+                            Toast.makeText(this, "Все данные загружены", Toast.LENGTH_SHORT).show();
+                        }
+                    }
+                });
     }
 
-    @RequiresApi(api = Build.VERSION_CODES.M)
     private void createRecycler() {
         adapter = new NewsAdapter(list);
         recyclerView.setAdapter(adapter);
         recyclerView.addItemDecoration(new DividerItemDecoration(getApplicationContext(), DividerItemDecoration.VERTICAL));
         recyclerView.setLayoutManager(new LinearLayoutManager(this, RecyclerView.VERTICAL, false));
-//        recyclerView.setOnScrollChangeListener((view, i, i1, i2, i3) -> {
-//            pageSize = 10;
-//            page++;
-//            mViewModel.receiveData(pageSize, page);
-//        });
-
     }
 }
